+++ BossMod/Modules/Heavensward/DeepDungeon/PalaceOfTheDead/D190TheGodfather.cs
@@ -22,7 +22,7 @@ public enum AID : uint
 
 //TODO: Make the boss's hitbox show up potentially/maybe add 2 circle indicators to show how close the stun bomb needs to be, add indicators to where the stun bomb will spawn next, to allow pre-positioning
 //spawn locations for stun bomb are as follows: 1:(-288.626, -300.256) 2:(-297.465, -297.525) 3:(-288.837, -305.537) 4:(-309.132, -303.739) 5:(-298.355, -293.630) 6:(-301.954, -314.289) 7:(-299.119, -297.563)
-class BossAdds(BossModule module) : Components.AddsMulti(module, [(uint)OID.LavaBomb, (uint)OID.RemedyBomb]);
+class BossAdds(BossModule module) : Components.AddsMulti(module, [OID.LavaBomb, OID.RemedyBomb]);
 class Flashthoom(BossModule module) : Components.SelfTargetedAOEs(module, ActionID.MakeSpell(AID.Flashthoom), new AOEShapeCircle(7.2f))
 {
     public override void AddAIHints(int slot, Actor actor, PartyRolesConfig.Assignment assignment, AIHints hints)
